requirements.txt
flask==3.1.1
python-docx==1.1.0
gunicorn==21.2.0
tiktoken==0.6.0
openai==1.14.3
beautifulsoup4==4.12.3
